// src/app/api/shared/mailbox/property-search-response.interface.ts
export interface PropertySearchResponseInterface extends Array<{
  city: string; // city
  full_address: string; // full address
  line1: string; // address line 1
  line2: string; // address line 2
  line3: string; // address line 3
  name: string; // property title
  ref: string;
  reference: string; // property reference
}> {

}